requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2==2.9.5
Pillow==9.3.0
django-oauth-toolkit==2.2.0
loguru==0.6.0

